HeaderMap other headers accessors + ProtectedHeaders refactor
Before it was EmptyOrSerializedHeaders which doesn't give much info over
how it is to be used.

HeaderMap's other_headers now have wrapper getters/setters which handle
the special cases of the reserves 1-7 keys which previously had
accessors.

Other getters for some structs were added which will be helpful for
deserialization users.