CBOR tests + fix other tests + misc fixes
Tests for cbor module

Other tests were updated to the new CBOR structs as they were still
using cbor_event's structs.

Float creation was disabled as cbor_event actually panics when trying to
serialize despite deserialization working fine. I am not sure of the
reasoning behind it but ot play it safe creation was disabled and an
error is returned (instead of panicing in cbor_event).